Models.dart separated into InputEntry.dart and MatchResult.dart, woozy_results.dart exports MatchResult.dart
