+++ lib/woozy_results.dart
@@ -1,0 +1,3 @@
+///The result class of the fuzzy-search library if needed somewhere explicitly
+
+export 'src/MatchResult.dart';
