+++ pubspec.yaml
@@ -1,6 +1,6 @@
 name: woozy_search
 description: A super simple and lightweight client-side fuzzy-search library based on Levenshtein distance.
-version: 2.0.2
+version: 2.0.3
 homepage: https://github.com/IvoriApp/woozy-search
 
 environment:
